+++ src/main/CharacterSets.ts
@@ -15,6 +15,7 @@ export class CharacterSet {
   CONTRACTION_20: string;
   CONTRACTION_30: string;
   CONTRACTION_40: string;
+  CONTRACTION_50: string;
   useTraditional: boolean;
   useCapital: boolean;
 
@@ -35,14 +36,15 @@ export class CharacterSet {
     this.CONTRACTION_20 = contraction[0];
     this.CONTRACTION_30 = contraction[1];
     this.CONTRACTION_40 = contraction[2];
+    this.CONTRACTION_50 = contraction[3];
   }
 }
 
 const allCharacterSets = [
   //
-  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌', false, false),
+  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩', false, false),
   new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点', '念', false, true),
-  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌', true, false),
+  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩', true, false),
   new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點', '念', true, true),
 ];