+++ src/test/data-contraction-all.ts
@@ -1,5 +1,6 @@
 import {data as data_20} from './data-contraction-20';
 import {data as data_30} from './data-contraction-30';
 import {data as data_40} from './data-contraction-40';
+import {data as data_50} from './data-contraction-50';
 
-export const data = [...data_20, ...data_30, ...data_40];
+export const data = [...data_20, ...data_30, ...data_40, ...data_50];